refactor(automation): extract CLI entry point into implementer_cli (Refs #468)
Move _parse_args, _setup_logging, and main out of the 872-line
IssueImplementer module into a dedicated implementer_cli module, separating
the CLI/entry-point concern from the orchestration class (SRP). main()
resolves its patchable collaborators (gh_list_open_issues, get_repo_root,
CursesUI, IssueImplementer) through the implementer module so existing tests
that patch implementer.<dep> and the hephaestus-implement-issues console
script keep working unchanged. The three callables are re-exported from
implementer with explicit `as` aliases.

implementer.py drops ~170 lines. Adds the new module to the frozen coverage
omit-allowlist (it drives the live claude/gh loop) and updates the allowlist
and orchestration-smoke guards accordingly.

This is one slice of the multi-PR #468 decomposition; the class's remaining
responsibilities were already extracted (StateManager, PhaseRunner,
WorktreeManager, pr_manager, SummaryPrinter), so #468 stays open for any
further narrowing.

Refs #468
