+++ hephaestus/automation/implementer.py
@@ -9,7 +9,6 @@ Provides:
 
 from __future__ import annotations
 
-import argparse
 import logging
 import subprocess
 import sys
@@ -20,10 +19,8 @@ from pathlib import Path
 from typing import Any
 
 from hephaestus.agents.runtime import (
-    add_agent_argument,
     is_codex,
 )
-from hephaestus.cli.utils import add_json_arg, emit_json_status
 
 # NOTE: Several symbols below are re-imported here purely so existing tests
 # can keep patching them at the ``hephaestus.automation.implementer.X`` path
@@ -38,13 +35,49 @@ from ._review_utils import find_pr_for_issue  # noqa: F401
 from .claude_invoke import invoke_claude_with_session  # noqa: F401
 from .curses_ui import CursesUI, ThreadLogManager
 from .dependency_resolver import CyclicDependencyError, DependencyResolver
-from .git_utils import get_repo_root, get_repo_slug, run  # noqa: F401
-from .github_api import fetch_issue_info, gh_list_open_issues
+
+# ``get_repo_root`` is re-exported with an explicit ``as`` alias so that
+# ``implementer_cli.main`` (which resolves it via this module) and tests
+# patching ``implementer.get_repo_root`` share one lookup site.
+from .git_utils import (  # get_repo_slug/run kept for test-patch compat
+    get_repo_root as get_repo_root,
+)
+from .git_utils import (
+    get_repo_slug,  # noqa: F401
+    run,
+)
+
+# ``gh_list_open_issues`` is re-exported with an explicit ``as`` alias (not
+# called in this module's body any more — ``main`` lives in ``implementer_cli``)
+# so tests can keep patching it at
+# ``hephaestus.automation.implementer.gh_list_open_issues`` and have
+# ``implementer_cli.main`` (which looks it up here) observe the patch.
+from .github_api import fetch_issue_info
+from .github_api import (
+    gh_list_open_issues as gh_list_open_issues,
+)
 
 # MAX_REVIEW_ITERATIONS is re-exported so tests that import it via
 # ``hephaestus.automation.implementer`` see the same value the runtime loop
 # in :class:`ImplementationPhaseRunner` uses. There is a single source of
 # truth in implementer_phase_runner.
+#
+# The CLI entry point (argument parsing, logging setup, and ``main``) lives in
+# ``implementer_cli`` (extracted for SRP — see #468). Re-exported here with
+# explicit ``as`` aliases (required by mypy) so the ``hephaestus-implement-issues``
+# console script (``hephaestus.automation.implementer:main``) keeps resolving and
+# existing tests calling ``implementer.main()`` / ``implementer._parse_args()``
+# and patching ``implementer.<dep>`` continue to work unchanged. ``main`` imports
+# this module lazily (inside its body), so this top-level import is cycle-safe.
+from .implementer_cli import (
+    _parse_args as _parse_args,
+)
+from .implementer_cli import (
+    _setup_logging as _setup_logging,
+)
+from .implementer_cli import (
+    main as main,
+)
 from .implementer_phase_runner import MAX_REVIEW_ITERATIONS, ImplementationPhaseRunner
 from .implementer_state import ImplementationStateManager
 from .implementer_summary import ImplementationSummaryPrinter
@@ -663,209 +696,6 @@ class IssueImplementer:
     def _print_summary(self, results: dict[int, WorkerResult]) -> None:
         """Print implementation summary."""
         ImplementationSummaryPrinter(self.worktree_manager).print(results)
-
-
-def _setup_logging(verbose: bool = False, log_dir: Path | None = None) -> None:
-    """Configure logging for the CLI.
-
-    Args:
-        verbose: Enable verbose (DEBUG) logging
-        log_dir: Optional directory to write log files
-
-    """
-    level = logging.DEBUG if verbose else logging.INFO
-    fmt = "%(asctime)s [%(levelname)s] %(name)s: %(message)s"
-    datefmt = "%Y-%m-%d %H:%M:%S"
-    logging.basicConfig(level=level, format=fmt, datefmt=datefmt)
-
-    if log_dir:
-        log_dir.mkdir(parents=True, exist_ok=True)
-        fh = logging.FileHandler(log_dir / "run.log", mode="a")
-        fh.setLevel(logging.DEBUG)
-        fh.setFormatter(logging.Formatter(fmt, datefmt=datefmt))
-        logging.getLogger().addHandler(fh)
-
-
-def _parse_args() -> argparse.Namespace:
-    """Parse command line arguments for the implementer CLI."""
-    parser = argparse.ArgumentParser(
-        description="Bulk implement GitHub issues using Claude Code",
-        formatter_class=argparse.RawDescriptionHelpFormatter,
-        epilog="""
-Examples:
-  # Implement all open issues (no arguments needed)
-  %(prog)s
-
-  # Implement all issues in an epic
-  %(prog)s --epic 123
-
-  # Implement specific issues
-  %(prog)s --issues 595 596 597
-
-  # Analyze dependencies without implementing
-  %(prog)s --epic 123 --analyze
-
-  # Resume previous implementation
-  %(prog)s --epic 123 --resume
-
-  # Health check
-  %(prog)s --health-check
-
-  # Dry run
-  %(prog)s --issues 595 --dry-run
-        """,
-    )
-
-    parser.add_argument(
-        "--epic",
-        type=int,
-        help="Epic issue number containing sub-issues",
-    )
-    parser.add_argument(
-        "--issues",
-        type=int,
-        nargs="+",
-        help="Specific issue numbers to implement (alternative to --epic)",
-    )
-    add_agent_argument(parser)
-    parser.add_argument(
-        "--analyze",
-        action="store_true",
-        help="Analyze dependencies without implementing",
-    )
-    parser.add_argument(
-        "--health-check",
-        action="store_true",
-        help="Run health check of dependencies and environment",
-    )
-    parser.add_argument(
-        "--resume",
-        action="store_true",
-        help="Resume previous implementation from saved state",
-    )
-    parser.add_argument(
-        "--max-workers",
-        type=int,
-        default=3,
-        choices=range(1, 33),
-        metavar="N",
-        help="Maximum number of parallel workers, 1-32 (default: 3)",
-    )
-    parser.add_argument(
-        "--no-skip-closed",
-        action="store_true",
-        help="Implement closed issues (default: skip closed issues)",
-    )
-    parser.add_argument(
-        "--no-auto-merge",
-        action="store_true",
-        help="Don't enable auto-merge on created PRs",
-    )
-    parser.add_argument(
-        "--dry-run",
-        action="store_true",
-        help="Show what would be done without actually doing it",
-    )
-    parser.add_argument(
-        "--no-learn",
-        action="store_true",
-        help="Disable /learn after implementation (enabled by default)",
-    )
-    parser.add_argument(
-        "--no-follow-up",
-        action="store_true",
-        help="Disable automatic filing of follow-up issues (enabled by default)",
-    )
-    parser.add_argument(
-        "--no-ui",
-        action="store_true",
-        help="Disable curses UI (use plain logging instead)",
-    )
-    parser.add_argument(
-        "-v",
-        "--verbose",
-        action="store_true",
-        help="Enable verbose logging",
-    )
-    add_json_arg(parser)
-
-    args = parser.parse_args()
-
-    if args.epic and args.issues:
-        parser.error("Cannot specify both --epic and --issues")
-
-    return args
-
-
-def main() -> int:
-    """Execute the issue implementation workflow.
-
-    Returns:
-        Exit code: 0 on success, 1 on failure, 130 on keyboard interrupt
-
-    """
-    args = _parse_args()
-
-    state_dir = get_repo_root() / "build" / ".issue_implementer"
-    _setup_logging(args.verbose, log_dir=state_dir)
-
-    log = logging.getLogger(__name__)
-
-    # Auto-discover all open issues when neither --issues nor --epic is given
-    if not args.health_check and not args.epic and not args.issues:
-        discovered = gh_list_open_issues()
-        log.info(
-            "No --issues/--epic given; discovered %s open issues: %s", len(discovered), discovered
-        )
-        args.issues = discovered
-
-    options = ImplementerOptions(
-        epic_number=args.epic or 0,
-        issues=args.issues or [],
-        agent=args.agent,
-        analyze_only=args.analyze,
-        health_check=args.health_check,
-        resume=args.resume,
-        max_workers=args.max_workers,
-        skip_closed=not args.no_skip_closed,
-        auto_merge=not args.no_auto_merge,
-        dry_run=args.dry_run,
-        enable_learn=not args.no_learn,
-        enable_follow_up=not args.no_follow_up,
-        enable_ui=not args.no_ui and not args.json,
-    )
-
-    if args.health_check:
-        log.info("Running health check")
-    elif args.issues:
-        log.info("Starting implementation of issues: %s", args.issues)
-    else:
-        log.info("Starting implementation of epic #%s", args.epic)
-
-    from hephaestus.utils.terminal import terminal_guard
-
-    with terminal_guard():
-        try:
-            implementer = IssueImplementer(options)
-            results = implementer.run()
-
-            if not args.health_check and not args.analyze:
-                failed = [num for num, result in results.items() if not result.success]
-                if failed:
-                    log.error("Failed to implement %s issue(s): %s", len(failed), failed)
-                    if args.json:
-                        emit_json_status(1, issues=args.issues or [], failed=failed)
-                    return 1
-
-            log.info("Complete")
-            if args.json:
-                emit_json_status(0, issues=args.issues or [], epic=args.epic or 0)
-            return 0
-        except KeyboardInterrupt:
-            log.warning("Interrupted by user")
-            if args.json:
-                emit_json_status(130, message="interrupted")
-            return 130
 
 
 if __name__ == "__main__":
