+++ tests/unit/validation/test_omit_allowlist.py
@@ -32,13 +32,14 @@ class TestOmitAllowlist:
 
         omit_list = pyproject.get("tool", {}).get("coverage", {}).get("run", {}).get("omit", [])
 
-        # Expected omit list: test globs + 10 automation modules
+        # Expected omit list: test globs + 11 automation modules
         expected_globs = {
             "*/tests/*",
             "*/__init__.py",
         }
         expected_modules = {
             "hephaestus/automation/implementer.py",
+            "hephaestus/automation/implementer_cli.py",
             "hephaestus/automation/implementer_phase_runner.py",
             "hephaestus/automation/implementer_summary.py",
             "hephaestus/automation/planner.py",
